Create New Claim
The create new claim method is now working. A claim can be made and
inserted into the claims table of the database. There is logic included
to ensure that a claim is able to be made. A claim can be made if the
user has been active for at least 6 months and has made less than 2
claims this year.